move battle_inferencer_for_distar, battle_rolloutor_for_distar to functional.collector.py
