+++ ding/framework/middleware/tests/test_league_pipeline.py
@@ -7,13 +7,12 @@ from ding.envs import BaseEnvManager, EnvSupervisor
 from ding.framework.supervisor import ChildType
 from ding.framework.context import BattleContext
 from ding.framework.middleware import StepLeagueActor, LeagueCoordinator, LeagueLearnerCommunicator, data_pusher, OffPolicyLearner
-from ding.framework.middleware.tests.mock_for_test import DIStarMockPolicy, DIStarMockPolicyCollect, battle_inferencer_for_distar, battle_rolloutor_for_distar
+from ding.framework.middleware.tests.mock_for_test import DIStarMockPolicy, DIStarMockPolicyCollect
+from ding.framework.middleware.functional.collector import battle_inferencer_for_distar, battle_rolloutor_for_distar
 from ding.framework.task import task, Parallel
 from ding.league.v2 import BaseLeague
 from dizoo.distar.config import distar_cfg
 from dizoo.distar.envs.distar_env import DIStarEnv
-from ding.framework.middleware.tests.mock_for_test import DIStarMockPolicy, DIStarMockPolicyCollect, \
-    battle_inferencer_for_distar, battle_rolloutor_for_distar
 from unittest.mock import patch
 
 env_cfg = dict(
