+++ ding/framework/middleware/tests/test_league_pipeline_one_process.py
@@ -2,15 +2,15 @@ from copy import deepcopy
 import pytest
 from ding.envs import BaseEnvManager, EnvSupervisor
 from ding.framework.context import BattleContext
-from ding.framework.middleware import StepLeagueActor, LeagueCoordinator, LeagueLearner
+from ding.framework.middleware import StepLeagueActor, LeagueCoordinator
 from ding.framework.supervisor import ChildType
 
 from ding.model import VAC
 from ding.framework.task import task
 from ding.framework.middleware.tests import cfg, MockLeague, MockLogger
 from dizoo.distar.envs.distar_env import DIStarEnv
-from ding.framework.middleware.tests import DIStarMockPolicy, DIStarMockPolicyCollect, \
-    battle_inferencer_for_distar, battle_rolloutor_for_distar
+from ding.framework.middleware.tests import DIStarMockPolicy, DIStarMockPolicyCollect
+from ding.framework.middleware.functional.collector import battle_inferencer_for_distar, battle_rolloutor_for_distar
 from distar.ctools.utils import read_config
 from unittest.mock import patch
 import os
